Refactored math functions into separate file
